// src/components/SlateEditor/types/element.ts
export enum ElementType {
  /**
   * Paragraph
   */
  PARAGRAPH = "paragraph",
  /**
   * @description 标题/正文
   */
  HEADER = "header",
  /**
   * @description 字体大小
   */
  FONT_SIZE = "fontSize",
  /**
   * @description 字体加粗
   */
  BOLD = "bold",
  /**
   * @description 字体斜体
   */
  ITALIC = "italic",
  /**
   * @description 字体下划线
   */
  UNDERLINE = "underline",
  /**
   * @description 字体删除线
   */
  STRIKETHROUGH = "strikethrough",
  /**
   * @description 引用
   */
  BLOCKQUOTE = "blockquote",
  /**
   * @description 链接
   */
  LINK = "link",
  /**
   * @description list-item
   */
  LIST_ITEM = "list-item",
  /**
   * @description 无序列表
   */
  OL = "orderedList",
  /**
   * @description 有序列表
   */
  UL = "unorderedList",
  /**
   * @description 字体颜色
   */
  FONT_COLOR = "fontColor",
  /**
   * @description 字体背景颜色
   */
  FONT_BG_COLOR = "bgColor",
  /**
   * @description 居左
   */
  ALIGN_LEFT = "alignLeft",
  /**
   * @description 居中
   */
  ALIGN_CENTER = "alignCenter",
  /**
   * @description 居右
   */
  ALIGN_RIGHT = "alignRight",
  /**
   * @description 上标
   */
  SUPERSCRIPT = "superscript",
  /**
   * @description 下标
   */
  SUBSCRIPT = "subscript",
  /**
   * @description 表格
   */
  TABLE = "table",
  /**
   * @description 表格中
   */
  IN_TABLE = "in-table",
  /**
   * @description 图片
   */
  IMAGE = "image",

  
}